# lib/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SoulDev_API",
    "description": "OpenAPI specification Test in GitLab",
    "contact": {
      "name": "Contact ",
      "url": "https://www.facebook.com/huynguyen07080401/",
      "email": "huynguyen07080401@gmail.com"
    },
    "license": {
      "name": "proprietary"
    },
    "version": "1.0"
  },
...
  "servers": [
    {
      "url": "http://localhost:4000/api",
      "description": "Test Server"
    },
    {
      "url": "http://3.26.183.207:4000/api",
      "description": "The production API server",
      "variables": {
        "username": {
          "default": "Production",
          "description": "this value is assigned by the service provider, in this example `gigantic-server.com`"
        },
        "port": {
          "enum": ["4040", "443"],
          "default": "4040"
        },
        "basePath": {
          "default": "v2"
        }
      }
    }
  ],
  "paths": {
...
    "/post/{id}": {
      "get": {
        "summary": "Get  1 Post",
        "description": "Some REST operation.",
        "security": [],
        "tags": ["Posts"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "number of items to skip",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the request was successful."
          },
          "400": {
            "description": "the request failed."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}